Gets traditional gates ontop of Qubits working
I managed to make AND, OR, and XOR gates using Qubits. The same functions can also take two bit inputs instead and do what classical gates do. Included in this commit is the incorperation of the bit_decoder package into SomeFramework and a new function that sets Qubits in a state in which you can use them as traditional bits for gates