Updates the README.md
I figured it was about time to update the README.md a bit. Now it has a bit better of an explination. As well, in this commit, I worked on determining what a Troffoli Gate can actually do and have derived a formula to determine its output when superpostioned around the Theta angle.